+++ ocdev/plugins/errors.py
@@ -1,0 +1,9 @@
+
+
+class DependencyError(Exception):
+
+    def __init__(self, msg):
+        self.msg = msg
+
+    def __str__(self):
+        return self.msg